# SwaggerUI.yml
swagger: "2.0"
info:
  description: "This is a documentation for Tours server.  The Tours server includes CRUD operations for tours ."
  version: "1.0.0"
  title: "Tours"
  contact:
    email: "ritesh.kamble@ebasolutions.net"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/api/v1/"
tags:
  - name: "tour"
    description: "Everything about the Tours"
  - name: "user"
    description: "Everything about the Users"
schemes:
  - "http"
paths:
  /tour:
    post:
      tags:
        - "tour"
      summary: "Crete a new tour"
      description: ""
      operationId: "createTour"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/tourCreate"
      responses:
        "405":
          description: "Invalid input"
    get:
      tags:
        - "tour"
      summary: "Get all tours"
      description: "Get all tours"
      produces:
        - "application/json"
      responses:
        "400":
          description: "Invalid request"
        "200":
          description: "Success"

  /tour/{id}:
    get:
      tags:
        - "tour"
      summary: "Finds tour by id"
      description: "Give single id in tour url"
      operationId: "getTour"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: path
          description: "Put tour Id's in number"
          type: "integer"
          required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "integer"
        "400":
          description: "Invalid status value"
    patch:
      tags:
        - "tour"
      summary: "Update Tour by id"
      description: "Tour id will be part of URL."
      operationId: "updateTour"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Tags to filter by"
          required: true
          type: "integer"
          items:
            type: "integer"
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/Tour"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "integer"
        "400":
          description: "Invalid ID "
    delete:
      tags:
        - "tour"
      summary: "Delete Tour by id"
      description: "Tour id will be part of URL."
      operationId: "deleteTour"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "Tags to filter by"
          required: true
          type: integer
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "integer"
        "400":
          description: "Invalid ID"
  /user:
    post:
      tags:
        - "user"
      summary: "Create user"
      description: "Create new user."
      operationId: "createUser"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        default:
          description: "successful operation"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
  Tour:
    type: "object"
    properties:
      tourName:
        type: "string"
  tourCreate:
    type: "object"
    properties:
      tourId:
        type: "integer"
      tourName:
        type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"